backend/requirements.txt
Django==4.2.8
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
Pillow==10.1.0
python-dotenv==1.0.0
django-filter==23.5
djangorestframework-simplejwt==5.5.1
celery==5.3.4
redis==5.0.1
shapely==2.0.2
fiona==1.9.5
openpyxl==3.1.2
requests==2.31.0
PyMySQL==1.1.0
reportlab==4.0.9
django-extensions==3.2.3
drf-yasg==1.21.7
pytest==7.4.3
pytest-django==4.7.0

# Production WSGI server
gunicorn==21.2.0
whitenoise==6.6.0